backend/requirements.txt
Flask==3.1.0
Werkzeug==3.1.3
pytest==8.0.0
pytest-cov==4.1.0
python-dotenv==1.0.0
requests==2.31.0
Flask-SQLAlchemy==3.1.1
psycopg2-binary==2.9.9
uuid7==0.1.0
Flask-Migrate==4.0.5
